+++ ChaosMod/ChaosMod.vcxproj
@@ -94,6 +94,7 @@
     <ClCompile Include="Effects\db\Misc\MiscEverythingFacePlayer.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscFireworks.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscForcedFP.cpp" />
+    <ClCompile Include="Effects\db\Misc\MiscForcefield.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscGravityController.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscHeatVision.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscIntenseMusic.cpp" />
@@ -119,7 +120,6 @@
     <ClCompile Include="Effects\db\Peds\PedsIntoRandomVehs.cpp" />
     <ClCompile Include="Effects\db\Player\PlayerHeavyRecoil.cpp" />
     <ClCompile Include="Effects\db\Player\PlayerWantedController.cpp" />
-    <ClCompile Include="Effects\db\Vehs\VehsDrivingPhobia.cpp" />
     <ClCompile Include="Effects\db\Weather\WeatherSnow.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscSpawnController.cpp" />
     <ClCompile Include="Effects\db\Misc\MiscTimecycModifierController.cpp" />
