+++ ChaosMod/ChaosMod.vcxproj.filters
@@ -156,7 +156,7 @@
     <ClCompile Include="Util\PoolSpawner.cpp" />
     <ClCompile Include="Effects\db\Vehs\VehsGTAOTraffic.cpp" />
     <ClCompile Include="Effects\db\Player\PlayerHeavyRecoil.cpp" />
-    <ClCompile Include="Effects\db\Vehs\VehsDrivingPhobia.cpp" />
+    <ClCompile Include="Effects\db\Misc\MiscForcefield.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DebugMenu.h" />
