+++ ConfigApp/Effects.cs
@@ -239,7 +239,7 @@ namespace ConfigApp
             EFFECT_PEDS_INTORANDOMVEHS,
             EFFECT_PLAYER_HEAVY_RECOIL,
             EFFECT_PEDS_CAT_GUNS,
-            EFFECT_VEH_DRIVING_PHOBIA,
+            EFFECT_MISC_FORCEFIELD,
             EFFECT_MISC_OIL_LEAKS,
             EFFECT_PEDS_FOG_OF_WAR,
             _EFFECT_ENUM_MAX
@@ -446,7 +446,7 @@ namespace ConfigApp
             {EffectType.EFFECT_PEDS_INTORANDOMVEHS, new EffectInfo("Set Everyone Into Random Vehicles", EffectCategory.PEDS, "peds_intorandomvehs")},
             {EffectType.EFFECT_PLAYER_HEAVY_RECOIL, new EffectInfo("Heavy Recoil", EffectCategory.PLAYER, "player_heavyrecoil", true)},
             {EffectType.EFFECT_PEDS_CAT_GUNS, new EffectInfo("Catto Guns", EffectCategory.PEDS, "peds_catguns", true)},
-            {EffectType.EFFECT_VEH_DRIVING_PHOBIA, new EffectInfo("Driving phobia", EffectCategory.VEHICLE, "veh_drivingphobia", true)},
+            {EffectType.EFFECT_MISC_FORCEFIELD, new EffectInfo("Forcefield", EffectCategory.MISC, "misc_forcefield", true)},
             {EffectType.EFFECT_MISC_OIL_LEAKS, new EffectInfo("Oil Leaks", EffectCategory.MISC, "misc_oilleaks", true)},
             {EffectType.EFFECT_PEDS_FOG_OF_WAR, new EffectInfo("Fog Of War", EffectCategory.PEDS, "peds_fogofwar", true)},
         };
